ath10k: new debugfs interface to enable adaptive CCA
For devices working in 2.4 GHz to be ETSI compliant,
it is required to support adaptive CCA. Some firmwares
support adaptive CCA and is disabled by default.

The patch is an attempt to add a new debugfs interface to enable
adaptive cca on supported firmwares. Support for this feature is
advertised with a new firmware feature flag.

Signed-off-by: Manikanta Pubbisetty <c_mpubbi@qti.qualcomm.com>