ath10k:  Ensure bcast, mcast, mgt rates are within bounds.
User may set a specific legacy rateset, and that may disable
the default bcast, mcast, and mgt rate values that the firmware
uses.

For that matter, 10.1 firmware has comments that indicate it should
send mgt frames at 1Mbps, but instead it at least sends the first
probe at 6Mbps.

So, choose the best rate based on rate-set and apply that to firmware.

If user has manually set the mcast, bcast, or mgt rate, then
use that regardless.  We will assume the user actually wants
the ucast rateset to be different than the mcast, bcast and/or mgt
rate.

Signed-off-by: Ben Greear <greearb@candelatech.com>